+++ packages/better_player/lib/src/configuration/player_event_type.dart
@@ -25,4 +25,5 @@ enum PlayerEventType {
   bufferingUpdate,
   bufferingEnd,
   changedPlaylistItem,
+  metadataReady,
 }